Fix another compiler warning on windows

+++ src/windows/winpthreads.h
@@ -794,7 +794,7 @@ static int pthread_create(pthread_t *th, pthread_attr_t *attr, void *(* func)(vo
 	if (attr)
 	{
 		tv->p_state = attr->p_state;
-		ssize = attr->s_size;
+		ssize = (unsigned int)attr->s_size;
 	}
 	
 	/* Make sure tv->h has value of -1 */
